Fixing PDF FE
