+++ client/src/App.css
@@ -17,6 +17,12 @@
   }
 }
 
+*,
+*::before,
+*::after {
+  border-style: none;
+}
+
 .App-header {
   background-color: #282c34;
   min-height: 100vh;
